requirements.txt
# Existing packages...
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
python-multipart==0.0.6
llama-parse==0.4.0
llama-index-core==0.10.11
nest_asyncio==1.6.0
chromadb==0.4.18
openai==1.54.0
python-dotenv==1.0.0
pydantic==2.5.0

# Production additions
gunicorn==21.2.0